+++ package.json
@@ -10,6 +10,7 @@
     "moment": "^2.24.0",
     "react": "^16.8.6",
     "react-dom": "^16.8.6",
+    "react-router-dom": "^5.2.0",
     "react-scripts": "^3.4.3",
     "react-spinner-material": "^1.1.1",
     "styled-components": "^4.2.1"
