+++ public/index.html
@@ -22,6 +22,9 @@
     -->
     <title>React App</title>
   </head>
+  <style>
+    @import url('https://fonts.googleapis.com/css2?family=Dancing+Script:wght@500&display=swap');
+</style>
   <body>
     <noscript>You need to enable JavaScript to run this app.</noscript>
     <div id="root"></div>
